+++ booking/src/test/java/rice_monkey/booking/service/BookingServiceTest.java
@@ -20,8 +20,8 @@ import rice_monkey.booking.exception.business.Authorization.UnauthorizedBookingA
 import rice_monkey.booking.exception.business.booking.AlreadyBookedException;
 import rice_monkey.booking.exception.business.booking.BookingNotFoundException;
 import rice_monkey.booking.exception.business.booking.ListingUnavailableException;
-import rice_monkey.booking.feign.ListingClient;
-import rice_monkey.booking.feign.dto.ListingDto;
+import rice_monkey.booking.feign.listing.ListingClient;
+import rice_monkey.booking.feign.listing.dto.ListingDto;
 
 import java.time.LocalDate;
 import java.util.Optional;
